refactor: extract tax advantaged create workflow
